Compiler warnings
git-svn-id: https://svn.code.sf.net/p/tcnopen/trdp/trunk@2206 3b5a3598-5f4e-4449-9e63-bd40438bfec0

+++ trdp/test/localtest/api_test.c
@@ -1911,11 +1911,11 @@ static int test14 ()
         /*
          Enter the main processing loop.
          */
-        int counter = 0;
+        unsigned int counter = 0;
         while (counter < 5)         /* 0.5 seconds */
         {
 
-            sprintf(data1, "Just a Counter: %08d", counter++);
+            sprintf(data1, "Just a Counter: %08u", counter++);
 
             err = tlp_put(gSession1.appHandle, pubHandle, (UINT8 *) data1, (UINT32) strlen(data1));
             IF_ERROR("tap_put");
@@ -2190,7 +2190,7 @@ static int test17 ()
 
         {
             UINT8   str[] = "123456789";
-            UINT32 result, seed, len = strlen((char*) str);
+            UINT32 result, seed, len = (UINT32) strlen((char*) str);
             /* CRC of the string "123456789" is 0x1697d06a ??? */
             seed = 0;
             result = vos_sc32(seed, str, len);
@@ -2198,7 +2198,7 @@ static int test17 ()
         }
         {
             UINT8   str[] = "123456789";
-            UINT32 result, seed, len = strlen((char*) str);
+            UINT32 result, seed, len = (UINT32) strlen((char*) str);
             /* CRC of the string "123456789" is 0x1697d06a ??? */
             seed = 0xFFFFFFFF;
             result = vos_sc32(seed, str, len);
@@ -2236,7 +2236,7 @@ static int test18 ()
         TRDP_COM_PAR_T          *pComPar;
         UINT32                  numIfConfig;
         TRDP_IF_CONFIG_T        *pIfConfig;
-        int i;
+        unsigned int i;
 
         err = tau_prepareXmlMem (xmlBuffer, strlen(xmlBuffer), &docHnd);
         IF_ERROR("tau_prepareXmlMem");
